Sdk merge

+++ wavesplatform/src/main/java/com/wavesplatform/sdk/model/request/node/BurnTransaction.kt
@@ -28,8 +28,8 @@ class BurnTransaction(
     override fun toBytes(): ByteArray {
         return try {
             Bytes.concat(
-                    byteArrayOf(type.toByte()),
-                    byteArrayOf(version.toByte()),
+                    byteArrayOf(type),
+                    byteArrayOf(version),
                     byteArrayOf(chainId),
                     Base58.decode(senderPublicKey),
                     Base58.decode(assetId),
