+++ wavesplatform/src/main/java/com/wavesplatform/sdk/model/request/node/SetScriptTransaction.kt
@@ -44,8 +44,8 @@ class SetScriptTransaction(
     override fun toBytes(): ByteArray {
         return try {
             Bytes.concat(
-                byteArrayOf(type.toByte()),
-                byteArrayOf(version.toByte()),
+                byteArrayOf(type),
+                byteArrayOf(version),
                 byteArrayOf(chainId),
                 Base58.decode(senderPublicKey),
                 scriptBytes(script),
